consolidate softwareinventory
