+++ comp/system-probe/softwareinventory/impl/module.go
@@ -1,32 +1,20 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
+// Copyright 2025-present Datadog, Inc.
 
-package modules
+//go:build windows
+
+package impl
 
 import (
 	"net/http"
 
 	"github.com/DataDog/datadog-agent/comp/system-probe/types"
 	"github.com/DataDog/datadog-agent/pkg/inventory/software"
-	"github.com/DataDog/datadog-agent/pkg/system-probe/api/module"
-	"github.com/DataDog/datadog-agent/pkg/system-probe/config"
-	sysconfigtypes "github.com/DataDog/datadog-agent/pkg/system-probe/config/types"
 	"github.com/DataDog/datadog-agent/pkg/system-probe/utils"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
-
-// SoftwareInventory Factory
-var SoftwareInventory = &module.Factory{
-	Name:             config.SoftwareInventoryModule,
-	ConfigNamespaces: []string{"software_inventory"},
-	Fn: func(_ *sysconfigtypes.Config, _ module.FactoryDependencies) (module.Module, error) {
-		return &softwareInventoryModule{}, nil
-	},
-}
-
-var _ module.Module = &softwareInventoryModule{}
 
 type softwareInventoryModule struct {
 }
